+++ tests/integration/repositories/mod.rs
@@ -4,5 +4,6 @@ mod authorization_codes;
 mod email_2fa_codes;
 mod recovery_codes;
 mod registered_clients;
+mod role_grant;
 mod roles;
 mod two_factor;